move SegmentReceipt into segment.rs
